requirements.txt
Django~=5.0.6
djangorestframework
djangorestframework-xml
dateparser